requirements.txt
soundcard>=0.4.2
soundfile>=0.12.1
numpy>=1.24.0
faster-whisper>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
jinja2>=3.1.0
